# docs/runtime_api_openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ClassificationReviewPayload": {
        "properties": {
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ClassificationReviewPayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Governance Atlas Runtime",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/classification-recommendations": {
      "get": {
        "operationId": "api_list_classification_recommendations_api_classification_recommendations_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assetFqn",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assetfqn"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api List Classification Recommendations",
        "tags": [
          "classification"
        ]
      }
    },
...
    "/api/classification-recommendations/{recommendation_id}": {
      "get": {
        "operationId": "api_get_classification_recommendation_api_classification_recommendations__recommendation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Get Classification Recommendation",
        "tags": [
          "classification"
        ]
      }
    },
    "/api/classification-recommendations/{recommendation_id}/review": {
      "post": {
        "operationId": "api_review_classification_recommendation_api_classification_recommendations__recommendation_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "recommendation_id",
            "required": true,
            "schema": {
              "title": "Recommendation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassificationReviewPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Review Classification Recommendation",
        "tags": [
          "classification"
        ]
      }
    },
...
  }
}